+++ src/core/data.js
@@ -162,7 +162,13 @@ module.exports = function(context) {
 
         if (d.files) d.type = 'gist';
         var type = d.length ? d[d.length - 1].type : d.type;
-        switch(type) {
+        if (d.commit) type = 'commit';
+        switch (type) {
+            case 'commit':
+                data.set({
+                    source: d.content
+                });
+                break;
             case 'local':
                 data.set({
                     type: 'local',
@@ -219,7 +225,8 @@ module.exports = function(context) {
                         login: login,
                         repo: repo,
                         branch: branch,
-                        name: d.name
+                        name: d.name,
+                        sha: d.sha
                     },
                     map: d.content,
                     path: d.path,